requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
pyglet==2.0.15
pillow==10.3.0